% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpham-data.r
\docType{data}
\name{alpham}
\alias{alpham}
\title{alpham Dataset for wmpvaer}
\format{typeof(alpham)=="list"}
\usage{
data(alpham)
}
\description{
Dataset needed for the operation of  R function "oneway_mss".
}
\references{
none
}
\keyword{datasets}
