% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_exp-data.r
\docType{data}
\name{chem_exp}
\alias{chem_exp}
\title{chem_exp Dataset for wmpvaer}
\format{typeof(chem_exp)=="list"}
\usage{
data(chem_exp)
}
\description{
Dataset is of unknown origin. Appears to be from a chemistry experiment.  
There are 45 observations and  8 columns. The columns are entitled (in order) 
H2S, P,K,Ca,Mg,Na,Mn,Zn. Assume H2S is the independent variable, levels of hydrogen
sulfide applied. The 7 other columns appear to be dependent variables corresponding to 
measurements of Phosphorus,Potassium,Calcium,Magnesium,Sodium,Manganese and Zinc. The dataset
is used to demonstrate application of a 1-Way MANOVA.
}
\references{
origin of this dataset is unknown
}
\keyword{datasets}
