% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{computeWilks}
\alias{computeWilks}
\title{computeWilks takes a matrix of eignvalues and computes values of the Wilks statistic}
\usage{
computeWilks(omega)
}
\arguments{
\item{omega}{Eigenvalues of an (E-inv)*H matrix,MANOVA analysis, Wilks non-centrality
parameters}
}
\description{
computeWilks takes a matrix of eignvalues and computes values of the Wilks statistic
}
