% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{find_min_s_hat}
\alias{find_min_s_hat}
\title{find_min_s_hat finds minimum s_hat for bisection search,s such that First-Der(s)=log(Wilks)}
\usage{
find_min_s_hat(omega, n, m)
}
\arguments{
\item{omega}{The a vector of eigenvalues of the Wilks Non-Centrality Parameter corresponding 
to one independent variable.}

\item{n}{Error DF for MANOVA}

\item{m}{The hypothesis Df of the one-way MANOVA analysis considered}
}
\description{
find_min_s_hat finds minimum s_hat for bisection search,s such that First-Der(s)=log(Wilks)
}
