% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\docType{class}
\name{mlandr-class}
\alias{mlandr-class}
\alias{mlandr}
\title{mlandr  S4 object}
\arguments{
\item{e_df}{error DF for a MANOVA model, numeric vector , or 1x1 numeric matrix}

\item{m_df}{model DFs for a MANOVA model, a numeric vector,length >=1, or a matrix ncol=1 & nrow>=1,}

\item{eigs}{Eigenvalues of an (E-inv)*H matrix,MANOVA analysis, Wilks non-centrality parameter, matrix w. ncol>=2}
}
\description{
\code{mlandr} the S4 object takes a matrix of eignvalues, a vector of model DF values and a numeric value of error DF
       and returns Wilks statistics & BW power and p values
}
