% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{mlandrf}
\alias{mlandrf}
\title{mlandrf  mlandr S4 object R function wrapper}
\usage{
mlandrf(e_dfm, m_dfm, eigsm)
}
\arguments{
\item{e_df}{error DF for a MANOVA model, numeric vector , or 1x1 numeric matrix}

\item{m_df}{model DFs for a MANOVA model, a numeric vector,length >=1, or a matrix ncol=1 & nrow>=1,}

\item{eigs}{Eigenvalues of an (E-inv)*H matrix,MANOVA analysis, Wilks non-centrality parameter, matrix w. ncol>=2}
}
\value{
list with two elements (matricies)
}
\description{
\code{mlandrf} takes a matrix of eignvalues, a vector of model DF values and a numeric value of error DF
       and returns Wilks statistics & BW power and p values by wrapping the S4 object 
       wmpvaer::mlandr in an R function.
}
\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(doParallel);
    library(wmpvaer);
    library(methods);
    data(eigs_plastic);
    eigsm<-as.matrix(eigs_plastic)
    colnames(eigsm)<-c(" "," "," ");
    rownames(eigsm)<-c("rate","additive","rate:additive");
    e_dfm<-16;
    m_dfm<-c(1,1,1);
    mlandrf(e_dfm,m_dfm,eigsm);
}
\keyword{&}
\keyword{Multivariate}
\keyword{Techniques}
\keyword{multivariate}
