% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{oneway_mss}
\alias{oneway_mss}
\title{oneway_mss   Computes est (based on the Wilks statistic) of # repetitions and sample 
              estimates for 1-way MANOVA w. 2 to 7 dependent vars and from 2 to 12 
              levels of an independent var.
\code{oneway_mss} takes two parameters: a scalar indicating number of levels and 
                a vector of the eigenvalues of a 1-way MANOVA (E-Inv)*H matrix.}
\usage{
oneway_mss(n_level, omega)
}
\arguments{
\item{n_level}{Is a numeric scalar, =>2 & =<12}

\item{omega}{Is numeric vector,eigenvalues of a 1-way MANOVA (E-Inv)*H matrix, length=>2 & =<7}
}
\value{
(if repetitions<=20) printed results.
}
\description{
oneway_mss   Computes est (based on the Wilks statistic) of # repetitions and sample 
              estimates for 1-way MANOVA w. 2 to 7 dependent vars and from 2 to 12 
              levels of an independent var.
\code{oneway_mss} takes two parameters: a scalar indicating number of levels and 
                a vector of the eigenvalues of a 1-way MANOVA (E-Inv)*H matrix.
}
\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(wmpvaer);
    rate_eigs<-c(1.6187719, -9.273797e-17,  3.365044e-18);
    rate_n_level<-2;  
    oneway_mss(rate_n_level,rate_eigs);
}
\keyword{&}
\keyword{Multivariate}
\keyword{Techniques}
\keyword{multivariate}
