% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{p_find_s}
\alias{p_find_s}
\title{p_find_s performs a bisection search, finds an s such that First-Der(s)=log(Wilks)}
\usage{
p_find_s(wilks, n, m, omega)
}
\arguments{
\item{wilks}{The value of the Wilks statistic for one independent variable}

\item{n}{The error Df of the one-way MANOVA analysis considered}

\item{m}{The hypothesis Df of the one-way MANOVA analysis considered}

\item{omega}{The a vector of eigenvalues of the Wilks Non-Centrality Parameter corresponding 
to one independent variable.}
}
\description{
p_find_s performs a bisection search, finds an s such that First-Der(s)=log(Wilks)
}
