% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_find_sBI}
\alias{p_find_sBI}
\title{p_find_sBI performs a bisection search, finds an s such that First-Der(s)=log(lambdaBI)}
\usage{
p_find_sBI(lambdaBI, N, psqrd)
}
\arguments{
\item{lambdaBI}{The value of the Block Independence statistic}

\item{N}{the number of columns in the data matrix}

\item{psqrd}{The a vector of squared eigenvalues}
}
\description{
p_find_sBI performs a bisection search, finds an s such that First-Der(s)=log(lambdaBI)
}
