% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{p_find_sECM}
\alias{p_find_sECM}
\title{p_find_sECM performs a bisection search, finds an s such that FirstDer(s)=log(lambdaECM)}
\usage{
p_find_sECM(lambdaECM, N1, N2, sigma_eigv)
}
\arguments{
\item{lambdaECM}{The value of the ECM statistic.}

\item{N1}{Number of rows in matrix 1}

\item{N2}{Number of rows in matrix 2}

\item{sigma_eigv}{A vector of eigenvalues of (Sigma11)*(inv-Sigma22)}
}
\description{
p_find_sECM performs a bisection search, finds an s such that FirstDer(s)=log(lambdaECM)
}
