% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{p_kprs2}
\alias{p_kprs2}
\title{p_kprs2 computes the 2nd derivative as a function of s_hat, s found by the function p_find_s}
\usage{
p_kprs2(s_hat, m, n, omega)
}
\arguments{
\item{s_hat}{The value needed for the first derivative to equal log(Wilks), s found by p_find_s}

\item{m}{The hypothesis Df of the one-way MANOVA analysis considered}

\item{n}{The error Df of the one-way MANOVA analysis considered}

\item{omega}{A vector of eigenvalues of the Wilks Non-Centrality Parameter 
corresponding to one independent variable.}
}
\description{
p_kprs2 computes the 2nd derivative as a function of s_hat, s found by the function p_find_s
}
