% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_kprs2BI}
\alias{p_kprs2BI}
\title{p_kprs2BI computes the 2nd derivative as a function of s_hat, 
          s_hat found by the function p_find_sBI}
\usage{
p_kprs2BI(s_hat, N, psqrd)
}
\arguments{
\item{s_hat}{The value needed for the first derivative to equal log(lambdaBI),
found by p_find_s}

\item{N}{the number of columns in the data matrix}

\item{psqrd}{The a vector of squared eigenvalues}
}
\description{
p_kprs2BI computes the 2nd derivative as a function of s_hat, 
          s_hat found by the function p_find_sBI
}
