% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{p_kprs2ECM}
\alias{p_kprs2ECM}
\title{p_kprs2ECM computes the 2nd der as a function of s_hat,s_hat found by function p_find_sECM}
\usage{
p_kprs2ECM(s_hat, N1, N2, sigma_eigv)
}
\arguments{
\item{s_hat}{The value needed for the first derivative to equal log(lambdaECM),found by p_find_sECM}

\item{N1}{Number of rows in matrix 1}

\item{N2}{Number of rows in matrix 2}

\item{sigma_eigv}{A vector of eigenvalues of (Sigma11)*(inv-Sigma22)}
}
\description{
p_kprs2ECM computes the 2nd der as a function of s_hat,s_hat found by function p_find_sECM
}
