% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_kprsBI}
\alias{p_kprsBI}
\title{p_kprsBI computes the first derivative as a function of s using an analytical formula}
\usage{
p_kprsBI(s, N, psqrd)
}
\arguments{
\item{s}{The value needed for the first derivative to equal log(lambdaBI)}

\item{N}{Number of columns in a data matrix}

\item{psqrd}{A vector of squared eigenvalues}
}
\description{
p_kprsBI computes the first derivative as a function of s using an analytical formula
}
