% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{p_l_and_r}
\alias{p_l_and_r}
\title{p_l_and_r computes the Lugananni and Rice approx. of power}
\usage{
p_l_and_r(wilks, n, m, omega)
}
\arguments{
\item{wilks}{The value of the Wilks statistic for one independent variable}

\item{n}{The error Df of the one-way MANOVA analysis considered}

\item{m}{The hypothesis Df of the one-way MANOVA analysis considered}

\item{omega}{A vector of eigenvalues of the Wilks Non-Centrality Parameter 
corresponding to one independent variable.}
}
\description{
p_l_and_r computes the Lugananni and Rice approx. of power
}
