% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_l_and_rBI}
\alias{p_l_and_rBI}
\title{p_l_and_rBI computes the Lugannani and Rice approx. of power}
\usage{
p_l_and_rBI(lambdaBI, N, psqrd)
}
\arguments{
\item{lambdaBI}{The value of a block independence likelihood ratio statistic}

\item{N}{the number of columns in the data matrix}

\item{psqrd}{The a vector of squared eigenvalues}
}
\description{
p_l_and_rBI computes the Lugannani and Rice approx. of power
}
