% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{p_l_and_rECM}
\alias{p_l_and_rECM}
\title{p_l_and_rECM computes the Lugannani and Rice approx. of power}
\usage{
p_l_and_rECM(lambdaECM, N1, N2, sigma_eigv)
}
\arguments{
\item{lambdaECM}{The value of a equality of covariance likelihood ratio statistic}

\item{N1}{Number of rows in matrix 1}

\item{N2}{Number of rows in matrix 2}

\item{sigma_eigv}{A vector of eigenvalues of (Sigma11)*(inv-Sigma22);}
}
\description{
p_l_and_rECM computes the Lugannani and Rice approx. of power
}
