% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_lmglmBI}
\alias{p_lmglmBI}
\title{p_lmglmBI, computation of log(mgf) of the log(lambdaBI) statistic.}
\usage{
p_lmglmBI(s_hat, N, psqrd)
}
\arguments{
\item{s_hat, }{The value needed for the first derivative to equal log(lambdaBI), found by function p_find_sBI}

\item{N}{the number of columns in the data matrix}

\item{psqrd}{A vector of squared eigenvalues}
}
\description{
p_lmglmBI, computation of log(mgf) of the log(lambdaBI) statistic.
}
