% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{p_mgamma}
\alias{p_mgamma}
\title{p_mgamma, computation of a multivariate gamma function needed by p_lmglm}
\usage{
p_mgamma(p, a)
}
\arguments{
\item{p}{The size or length of a required eigenvector}

\item{a}{A variable computed in p_lmglm}
}
\description{
p_mgamma, computation of a multivariate gamma function needed by p_lmglm
}
