% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_mgammaBI}
\alias{p_mgammaBI}
\title{p_mgammaBI, computation of a multivariate gamma function needed by p_lmglmBI}
\usage{
p_mgammaBI(p, a)
}
\arguments{
\item{p}{The size or length of a required eigenvector}

\item{a}{A variable computed in p_lmglmBI}
}
\description{
p_mgammaBI, computation of a multivariate gamma function needed by p_lmglmBI
}
