% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{p_mgammaECM}
\alias{p_mgammaECM}
\title{p_mgammaECM, computation of a multivariate gamma function needed by p_lmglmECM}
\usage{
p_mgammaECM(p, a)
}
\arguments{
\item{p}{The size or length of a required eigenvector}

\item{a}{A variable computed in p_lmglmECM}
}
\description{
p_mgammaECM, computation of a multivariate gamma function needed by p_lmglmECM
}
