% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{p_oneFhatone}
\alias{p_oneFhatone}
\title{p_oneFhatone, computation of an approximate 1F1 value}
\usage{
p_oneFhatone(s, n, m, omega)
}
\arguments{
\item{s}{The value needed for the first derivative to equal log(Wilks)}

\item{n}{The error Df of the one-way MANOVA analysis considered}

\item{m}{The hypothesis Df of the one-way MANOVA analysis considered}

\item{omega}{The a vector of eigenvalues of the Wilks Non-Centrality Parameter 
corresponding to one independent variable.}
}
\description{
p_oneFhatone, computation of an approximate 1F1 value
}
