% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{p_twoF1BI}
\alias{p_twoF1BI}
\title{p_twoF1BI, computation of an approximate 2F1 value for block independence}
\usage{
p_twoF1BI(s, N, psqrd)
}
\arguments{
\item{s}{The value needed for the first derivative to equal log(lambdaBI)}

\item{N}{number of columns in the data matrix}

\item{psqrd}{The a vector of squared eigenvalues}
}
\description{
p_twoF1BI, computation of an approximate 2F1 value for block independence
}
