% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{p_twoF1ECM}
\alias{p_twoF1ECM}
\title{p_twoF1ECM, computation of an approximate 2F1 value for equality of covariances}
\usage{
p_twoF1ECM(s, N1, N2, sigma_eigv)
}
\arguments{
\item{s}{The value needed for the first derivative to equal log(lambdaECM)}

\item{N1}{number of rows in data matrix N1}

\item{N2}{number of rows in data matrix N2}

\item{sigma_eigv}{A vector of eigenvalues of (Sigma11)*(inv-Sigma22)}
}
\description{
p_twoF1ECM, computation of an approximate 2F1 value for equality of covariances
}
