% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plastic-data.R
\docType{data}
\name{plastic}
\alias{plastic}
\title{Plastic Film Dataset for wmpvaer}
\format{typeof(plastic)=="list"}
\usage{
data(plastic)
}
\description{
Data from an industrial experiment from the textbook of Johnson and Wichern,
1992. Data is also found in R documentation as an example of the function 
manova.The dependent variables are taken to be "tear","gloss" and "opacity". The
independent variables are taken to be "rate" (actually "rate of extrusian" orginally)
and "additive".
}
\references{
Johnson,R.A. and Wichern, D.W. 1992, Applied Multivariate Statistical Analysis
}
\keyword{datasets}
