% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils-data.R
\docType{data}
\name{soils}
\alias{soils}
\title{Soils Characteristics from Different Contours and Depths Dataset for wmpvaer}
\format{typeof(soils)=="list"}
\usage{
data(soils)
}
\description{
Data from an examination of soils data published in 1968. The independent variables are "block",
"contour" and "depth". All other variable (9 in total) are soil characteristic measurements. This
dataset is widely used in examples of MANOVA.
}
\references{
Horton,IF,Russel, JS, Moore,AW (1968) "Multivariate-Covariance and Cannonical Analysis:A Method for Selecting The 
  Most Effective Discriminators in a Multivariate Situation", Biometrics, 24(4), 845-858
}
\keyword{datasets}
