% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{statsBIf}
\alias{statsBIf}
\title{statsBIf   Computes lambdaBI statistics,p-values, and shows the block matrix definitions}
\usage{
statsBIf(data_matrix, block)
}
\arguments{
\item{data_matrix}{Is a numeric matrix}

\item{block}{Is a scalar or vector of column numbers corresponding to block 1}
}
\value{
list with one matrix as element
}
\description{
\code{statsBIf} takes a matrix of numeric data, and a vector or scalar indicating columns numbers 
               for a block and returns the LR statistic lambdaBI, computed p values and a matrix w. column names 
               and numbers
}
\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(wmpvaer);
    data(plastic);
    plasticm<-as.matrix(plastic[,1:3]);
    block_v<-c(2)  
    statsBIf(plasticm,block_v);
}
\keyword{&}
\keyword{Multivariate}
\keyword{Techniques}
\keyword{multivariate}
