% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_functions.r
\name{statsBIncatf}
\alias{statsBIncatf}
\title{statsBIncatf Computation with block1 columns taken in combinations n at a time}
\usage{
statsBIncatf(data_matrix, nc_at_time)
}
\arguments{
\item{data_matrix}{A data matrix composed of numeric values}

\item{nc_at_time}{A scalar that identifies the value of number of columns at a time"}
}
\value{
list statsBIf matracies
}
\description{
\code{statsBIncatf} Using the function statsBIf computes stats for block independence 
           using columns from data_matrix at specified "n columns at a time".The program
           uses parallel computation if possible.
}
\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(doParallel);
    library(wmpvaer);
    data(plastic);
    plasticm<-as.matrix(plastic[,1:3]);
    nc_at_time<-1;  
    statsBIncatf(plasticm,nc_at_time);       
}
