% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECM_functions.r
\name{statsECMf}
\alias{statsECMf}
\title{statsECMf   Computes a lambdaECM statistics and p-values}
\usage{
statsECMf(data_matrix1, data_matrix2)
}
\arguments{
\item{data_matrix1}{Is a numeric matrix,}

\item{data_matrix2}{Is a numeric matrix,}
}
\value{
One matrix with two elements
}
\description{
\code{statsECMf} takes two matricies of numeric data, the LR statistic lambdaECM, computed p values
}
\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(wmpvaer);
    data(plastic)
    a1<-plastic[plastic$additive=="Low",];
    low_additive<-as.matrix(a1[,1:3]);
    a2<-plastic[!(plastic$additive=="Low"),];
    high_additive<-as.matrix(a2[,1:3]);
    statsECMf(low_additive,high_additive);
}
\keyword{&}
\keyword{Multivariate}
\keyword{Techniques}
\keyword{multivariate}
