% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r
\name{valid_mlandr}
\alias{valid_mlandr}
\title{valid_mlandr contains validity tests for the wmpvaer::mlandr object}
\usage{
valid_mlandr(object)
}
\arguments{
\item{object}{The object in question is the S4 wmpvaer::mlandr}
}
\description{
valid_mlandr contains validity tests for the wmpvaer::mlandr object
}
