% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_and_functions.r, R/wmpvaer.r
\docType{package}
\name{wmpvaerf}
\alias{wmpvaerf}
\alias{wmpvaerf}
\alias{wmpvaerf-package}
\title{Wilks MANOVA p-Value Alternative Estimation}
\usage{
wmpvaerf(formula)
}
\arguments{
\item{formula}{Same formula as used with the R MANOVA function}
}
\value{
list with three elements (matricies)
}
\description{
\code{wmpvaerf} takes Wilks values from R's MANOVA function, adds power estimates and alternative p-values

The wmpvaer package provides four exported functions:
wmpvaerf, mlandrf,statsBIf,statBIncatf,statsECMf
}
\section{Warning}{

This code example,on a single core, can take 30 seconds or more to complete. The program exploits two or more cores
if available,however.On a three core Intel system the code example takes about 11 seconds.
}

\examples{
    suppressPackageStartupMessages(library(Rmpfr)); 
    library(doParallel);
    library(wmpvaer);
    data(plastic);
    Y<-as.matrix(plastic[,1:3]);
    rate<-matrix(data=(plastic$rate),nrow(plastic),1);
    additive<-matrix(data=(plastic$additive),nrow(plastic),1); 
    wmpvaerf(Y~ rate*additive);
}
\keyword{&}
\keyword{Multivariate}
\keyword{Techniques}
\keyword{multivariate}
